+++ ts/src/components/grid/types.ts
@@ -1,24 +1,17 @@
 import { PropsWithChildren, HTMLAttributes, ReactElement } from 'react';
+import { DataModel, DataStore } from '../../core';
 
 export interface GridProps extends PropsWithChildren<any> {
-  // data?: any[],
-  // className?: string,
-  // fitScreen?: boolean,
-  // fitWidth?: boolean,
-  // fitHeight?: boolean,
-  // footer?: boolean,
-  // footerRenderer?: (records: any[]) => ReactElement,
-  // checkboxSelection?: boolean,
-  // onCheckChange?: (record: any) => void,
-  // transpose?: boolean,
+  data?: any[],
+  store?: DataStore<any>,
+  checkboxSelection?: boolean,
+  onSelectionChange?: (record: DataModel<any>) => void,
+  bbar?: (records: DataModel<any>[]) => ReactElement,
+  pagingToolbar?: { pageSize?: number },
 }
 
 export interface GridColumnProps extends HTMLAttributes<HTMLDivElement> {
   field: string,
   headerName?: string,
-  // locked?: boolean,
-  // numeric?: boolean,
-  // numericFormat?: string,
-  // editable?: boolean,
-  // cellRenderer?: (value: any, item: any, rowIndex: number, colIndex: number, record: any) => string | number | ReactElement,
+  renderer?: (value: any, record?: DataModel<any>, rowIndex?: number, colIndex?: number) => string | number | ReactElement,
 }