+++ ts/src/core/index.ts
@@ -3,3 +3,6 @@ export * from './cache';
 export * from './ajax';
 export * from './observable';
 export * from './data';
+
+import * as _Rosie from './utils';
+export const Rosie = { ..._Rosie };