+++ ts/src/index.ts
@@ -5,6 +5,3 @@ import './lang/array';
 
 export * from './core';
 export * from './components';
-
-import * as _Rosie from './utils';
-export const Rosie = { ..._Rosie };