+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 ## Next
 
 - New Features
-  - #6 Add DataModel, DataStore object
+  - #6 Add `DataModel`, `DataStore` object
+- Improvements
+  - #9 Supports to select row by checkbox
+- Bug Fixes
+  - #13 `GridCell` not update value when `DataModel` change bug
 
 ## 0.1.3
 > 2024-01-13